gcp_vertex_ai/requirements.txt
flask
google-cloud-aiplatform
google-cloud-storage
gunicorn
Werkzeug